Git was reset due to incorrect creation of repository, previous history lost
This project is by a default Kotlin Project but I don't think it is a problem, Java classes can still be created.
This default project starts with navigation built in and we can build off of that. We should modify our code to work with the default code. That should be helpful.

I believe the database is working but I will check that soon.